+++ impl/oxford_dictionary.dart
@@ -3,8 +3,6 @@
 // All rights reserved
 
 import 'package:dictosaurus/dictosaurus.dart';
-// import 'package:dictosaurus/src/dictionary/dictionary.dart';
-import 'package:dictosaurus/type_definitions.dart';
 import 'package:gmconsult_dart_core/dart_core.dart';
 import 'dart:async';
 
@@ -116,10 +114,10 @@ extension _OxfordDictionariesHashmapExtension on Map<String, dynamic> {
     return retVal;
   }
 
-  /// Returns the `id` field of the JSON response as `String?`.
+  /// Returns the `id` field of the Map<String, dynamic> response as `String?`.
   String? get term => this['id'] is String ? this['id'] as String : null;
 
-  /// Returns the `language` field of the JSON response as `String?`.
+  /// Returns the `language` field of the Map<String, dynamic> response as `String?`.
   String? get language =>
       (this['language'] is String ? this['language'] as String : null)
           ?.replaceAll('-', '_');
@@ -139,9 +137,10 @@ extension _OxfordDictionariesHashmapExtension on Map<String, dynamic> {
     return null;
   }
 
-  Iterable<JSON> getJsonList(String fieldName) => this[fieldName] is Iterable
-      ? (this[fieldName] as Iterable).cast<JSON>()
-      : [];
+  Iterable<Map<String, dynamic>> getJsonList(String fieldName) =>
+      this[fieldName] is Iterable
+          ? (this[fieldName] as Iterable).cast<Map<String, dynamic>>()
+          : [];
 
   Iterable<String> getStringList(String fieldName) =>
       this[fieldName] is Iterable
@@ -151,7 +150,7 @@ extension _OxfordDictionariesHashmapExtension on Map<String, dynamic> {
   TermProperties? toTermProperties() {
     final term = this.term;
     String languageCode = '';
-    final Iterable<JSON> results = getJsonList('results');
+    final Iterable<Map<String, dynamic>> results = getJsonList('results');
     if (results.isNotEmpty && term is String) {
       languageCode =
           languageCode.isEmpty ? results.first.language ?? '' : languageCode;
@@ -257,7 +256,7 @@ enum OxFordDictionariesEndpoints {
   inflections
 }
 
-/// A mixin that returns [JSON] responses from the Oxford Dictionaries API.
+/// A mixin that returns [Map<String, dynamic>] responses from the Oxford Dictionaries API.
 ///
 /// See: https://developer.oxforddictionaries.com/.
 abstract class OxfordDictionariesApiMixin {
@@ -308,7 +307,7 @@ abstract class OxfordDictionariesApiMixin {
   ///
   /// More information at:
   /// https://developer.oxforddictionaries.com/documentation#!/Words/get_words_source_lang
-  Future<JSON?> entriesEndPoint(String term,
+  Future<Map<String, dynamic>?> entriesEndPoint(String term,
       {bool strictMatch = false, String sourceLanguage = 'en-us'}) async {
     final languages =
         kLanguagesMap[OxFordDictionariesEndpoints.entries] as Set<String>;