+++ lib/dictosaurus.dart
@@ -3,12 +3,10 @@
 // All rights reserved
 
 /// Dictionary, thesaurus and term expansion utilities.
-///
-/// Also exports the core classes from the `text_indexing` package.
 library dictosaurus;
 
 export 'src/auto_correct/auto_correct.dart' show AutoCorrect;
-export 'src/thesaurus/thesaurus.dart' show Thesaurus;
+export 'src/thesaurus/thesaurus.dart' show Thesaurus, ThesaurusMixin;
 export 'src/dictionary/_index.dart'
     show
         Dictionary,
@@ -19,4 +17,3 @@ export 'src/dictionary/_index.dart'
         PartOfSpeech;
 export 'src/dictosaurus.dart'
     show DictoSaurus, DictoSaurusBase, DictoSaurusMixin;
-export 'package:text_indexing/text_indexing.dart';