+++ lib/src/_index.dart
@@ -6,7 +6,4 @@ export 'auto_correct/auto_correct.dart';
 export 'thesaurus/thesaurus.dart';
 export 'dictionary/_index.dart';
 export 'dictosaurus.dart';
-export 'package:text_indexing/text_indexing.dart';
-export 'package:text_indexing/type_definitions.dart';
-export 'package:text_indexing/extensions.dart';
 export 'typedefs.dart';