+++ lib/type_definitions.dart
@@ -3,11 +3,6 @@
 // All rights reserved
 
 /// Typedefs used by the `dictosaurus` package.
-///
-/// Also includes the typedefs from the `text_indexing` package.
 library type_definitions;
 
-/// Export of barrel file in the /src folder.
-/// src/_index.dart exports the barrel files from
 export 'src/typedefs.dart';
-export 'package:text_indexing/type_definitions.dart';